datasets/SWE-Refactor/code/requirements.txt
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
codebleu==0.7.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
docker-pycreds==0.4.0
durationpy==0.7
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
javalang==0.13.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.1
langchain-anthropic==0.3.0
langchain-community==0.3.1
langchain-core==0.3.19
langchain-experimental==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.22
langgraph-checkpoint==1.0.10
langsmith==0.1.130
litellm==1.50.0
lxml==5.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.52.0
openpyxl==3.1.5
orjson==3.10.7
overrides==7.7.0
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
posthog==3.6.6
protobuf==4.25.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rerankers==0.5.3
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.1.1
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomli==2.0.1
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3



